move AdvancedLogger.h to minifi-api/include/minifi-cpp/core/logging/AdvancedLogger.h
