+++ extensions/standard-processors/processors/LogAttribute.h
@@ -25,14 +25,14 @@
 #include <vector>
 
 #include "core/Core.h"
-#include "core/ProcessorImpl.h"
-#include "minifi-cpp/core/ProcessContext.h"
 #include "core/ProcessSession.h"
-#include "minifi-cpp/core/PropertyDefinition.h"
+#include "core/ProcessorImpl.h"
 #include "core/PropertyDefinitionBuilder.h"
+#include "minifi-cpp/core/ProcessContext.h"
+#include "minifi-cpp/core/PropertyDefinition.h"
 #include "minifi-cpp/core/PropertyValidator.h"
 #include "minifi-cpp/core/RelationshipDefinition.h"
-#include "core/logging/LoggerFactory.h"
+#include "minifi-cpp/core/logging/AdvancedLogger.h"
 #include "minifi-cpp/utils/Export.h"
 
 namespace org::apache::nifi::minifi::processors {
